rename ft_math
