+++ float/ft_fmin.c
@@ -1,26 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libft.h                                            :+:      :+:    :+:   */
+/*   ft_fmin.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 14:54:41 by reasuke           #+#    #+#             */
-/*   Updated: 2024/03/20 22:46:12 by reasuke          ###   ########.fr       */
+/*   Created: 2023/11/30 19:53:04 by reasuke           #+#    #+#             */
+/*   Updated: 2023/12/18 17:47:53 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LIBFT_H
-# define LIBFT_H
+#include "libft.h"
 
-# include "ft_file.h"
-# include "ft_math.h"
-# include "ft_list.h"
-# include "ft_memory.h"
-# include "ft_output.h"
-# include "ft_printf.h"
-# include "ft_string.h"
-# include "ft_type.h"
-# include "get_next_line.h"
-
-#endif
+double	ft_fmin(double a, double b)
+{
+	if (b < a)
+		return (b);
+	return (a);
+}
