+++ float/ft_fswap.c
@@ -1,26 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libft.h                                            :+:      :+:    :+:   */
+/*   ft_fswap.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 14:54:41 by reasuke           #+#    #+#             */
-/*   Updated: 2024/03/20 22:46:12 by reasuke          ###   ########.fr       */
+/*   Created: 2023/11/30 20:29:07 by reasuke           #+#    #+#             */
+/*   Updated: 2023/12/18 17:48:03 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LIBFT_H
-# define LIBFT_H
+#include "libft.h"
 
-# include "ft_file.h"
-# include "ft_math.h"
-# include "ft_list.h"
-# include "ft_memory.h"
-# include "ft_output.h"
-# include "ft_printf.h"
-# include "ft_string.h"
-# include "ft_type.h"
-# include "get_next_line.h"
+void	ft_fswap(double *a, double *b)
+{
+	double	tmp;
 
-#endif
+	tmp = *a;
+	*a = *b;
+	*b = tmp;
+}
