+++ ft_math/ft_chmin.c
@@ -1,26 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libft.h                                            :+:      :+:    :+:   */
+/*   ft_chmin.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 14:54:41 by reasuke           #+#    #+#             */
-/*   Updated: 2024/03/20 22:46:12 by reasuke          ###   ########.fr       */
+/*   Created: 2023/11/30 19:54:01 by reasuke           #+#    #+#             */
+/*   Updated: 2024/03/20 22:47:49 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LIBFT_H
-# define LIBFT_H
+#include "ft_math.h"
 
-# include "ft_file.h"
-# include "ft_math.h"
-# include "ft_list.h"
-# include "ft_memory.h"
-# include "ft_output.h"
-# include "ft_printf.h"
-# include "ft_string.h"
-# include "ft_type.h"
-# include "get_next_line.h"
-
-#endif
+bool	ft_chmin(int *a, int b)
+{
+	if (b < *a)
+	{
+		*a = b;
+		return (true);
+	}
+	return (false);
+}
