+++ ft_math/ft_max.c
