+++ ft_math/ft_min.c
