+++ include/ft_math.h
@@ -1,26 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   libft.h                                            :+:      :+:    :+:   */
+/*   ft_math.h                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: reasuke <reasuke@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/09/02 14:54:41 by reasuke           #+#    #+#             */
-/*   Updated: 2024/03/20 22:46:12 by reasuke          ###   ########.fr       */
+/*   Created: 2024/02/24 12:09:36 by reasuke           #+#    #+#             */
+/*   Updated: 2024/03/20 22:47:01 by reasuke          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef LIBFT_H
-# define LIBFT_H
+#ifndef FT_MATH_H
+# define FT_MATH_H
 
-# include "ft_file.h"
-# include "ft_math.h"
-# include "ft_list.h"
-# include "ft_memory.h"
-# include "ft_output.h"
-# include "ft_printf.h"
-# include "ft_string.h"
-# include "ft_type.h"
-# include "get_next_line.h"
+# include <stdbool.h>
+
+int		ft_abs(int a);
+int		ft_max(int a, int b);
+int		ft_min(int a, int b);
+bool	ft_chmax(int *a, int b);
+bool	ft_chmin(int *a, int b);
+void	ft_swap(int *a, int *b);
 
 #endif
