+++ unit_test.mk
@@ -17,7 +17,7 @@ TEST_SRC		= $(TEST_DIR)/test_strcmp.cpp \
 					$(TEST_DIR)/test_strstr.cpp \
 					$(TEST_DIR)/test_strndup.cpp \
 					$(TEST_DIR)/test_strtol.cpp \
-					$(TEST_DIR)/test_integer.cpp \
+					$(TEST_DIR)/test_math.cpp \
 					$(TEST_DIR)/test_lst_before.cpp \
 					$(TEST_DIR)/test_file_to_lines.cpp
 TEST_OBJ		= $(patsubst $(TEST_DIR)/%.cpp, $(TEST_BUILD_DIR)/%.o, $(TEST_SRC))
